Description of the project so far.
Intent is to show that the project is now
somewhat 'rebased' in a new directory; and
at present is an exact copy of the existing
work in the parent directory.

THE LOST WORLD file ideas were followed, to
make the attempt to copy-with-history (so that
git will show the history of these copies as
if they were the originals).

	new file:   WHAT.THIS.IS.md

On branch dvlp-aa